combine test_tango_one_of_everything and test_tango_signals
Don't skip enums in tango test_tango_signal tests
